Split configuration and vo.
