+++ src/llm_chatbot_for_messengers/core/workflow/qa.py
@@ -12,17 +12,19 @@ from langchain_core.runnables import Runnable, RunnablePassthrough
 from langgraph.graph import END, START
 from typing_extensions import override
 
+from llm_chatbot_for_messengers.core.configuration import (
+    WorkflowNodeConfig,
+)
 from llm_chatbot_for_messengers.core.error import WorkflowError
 from llm_chatbot_for_messengers.core.output.template import get_template
-from llm_chatbot_for_messengers.core.vo import (
+from llm_chatbot_for_messengers.core.workflow.base import Workflow
+from llm_chatbot_for_messengers.core.workflow.vo import (
     AnswerNodeResponse,
     QAState,
     SummaryNodeDocument,
     SummaryNodeResponse,
     WebSummaryState,
-    WorkflowNodeConfig,
 )
-from llm_chatbot_for_messengers.core.workflow.base import Workflow
 
 if TYPE_CHECKING:
     from langchain_core.language_models import BaseChatModel
