+++ tests/core/test_vo.py
@@ -1,6 +1,7 @@
 import pytest
+from llm_chatbot_for_messengers.core.configuration import AgentConfig, AgentExtraConfig, LLMConfig, WorkflowNodeConfig
 from llm_chatbot_for_messengers.core.output.memory import VolatileMemoryManager
-from llm_chatbot_for_messengers.core.vo import AgentConfig, AgentExtraConfig, LLMConfig, UserId, WorkflowNodeConfig
+from llm_chatbot_for_messengers.core.vo import UserId
 from pydantic import ValidationError
